pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Morris Viitanen <morris.viitanen@helsinki.fi>"]
readme = "README.md"
packages = [{include = "ot_harjoitustyo"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
